r/aws_cloudtrail_organization_delegated_admin_account: Serialize acceptance tests.
